third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEDistantLightElement-svgdom-elevation-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'elevation' property of the SVGFEDistantLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS distantLight.elevation.baseVal is 50
PASS distantLight.elevation.baseVal is 20
PASS successfullyParsed is true

TEST COMPLETE
